requirements.txt
# FastAPI and ASGI server
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# AI/ML Models
openai==1.3.7
tiktoken==0.5.1

# Database
motor==3.3.2  # Async MongoDB driver
pymongo==4.6.0
redis==5.0.1

# Vector Database
pinecone-client==2.2.4
# Or use: supabase==2.0.3

# Scheduling
apscheduler==3.10.4
celery==5.3.4

# Audio Processing
pydub==0.25.1
soundfile==0.12.1
numpy==1.24.3

# HTTP Requests
httpx==0.25.2
requests==2.31.0

# Web Search
tavily-python==0.7.13

# Environment
python-dotenv==1.0.0

# Date/Time
python-dateutil==2.8.2
pytz==2023.3

# Validation
pydantic==2.5.0
pydantic-settings==2.1.0

# Authentication (optional)
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Logging
loguru==0.7.2

# Firebase for push notifications
firebase-admin==6.3.0
